Removed duplicate file structure to the root directory.
